+++ tests/translation_compare.py
@@ -1,96 +1,169 @@
+import asyncio
 import json
+import logging
 import os
+from pathlib import Path
+import sys
 
-# Requires running cmd: "pip install flatten_json"
-from flatten_json import flatten
+from flatten_json import flatten, unflatten
+import translators as ts
 
-from custom_components.mydolphin_plus import DOMAIN
+DOMAIN = "mydolphin_plus"
 
-SUPPORTED_LANGUAGES = ["en", "it"]
+DEBUG = str(os.environ.get("DEBUG", False)).lower() == str(True).lower()
 
+log_level = logging.DEBUG
 
-def _get_json_file(path):
-    full_path = os.path.join(
-        os.path.dirname(__file__),
-        f"..\\custom_components\\{DOMAIN}\\{path}"
-    )
+root = logging.getLogger()
+root.setLevel(log_level)
 
-    with open(full_path, encoding="utf-8") as json_file:
-        json_str = json_file.read()
-        content = json.loads(json_str)
+logging.getLogger("urllib3").setLevel(logging.WARNING)
 
-        return content
+stream_handler = logging.StreamHandler(sys.stdout)
+stream_handler.setLevel(log_level)
+formatter = logging.Formatter("%(asctime)s %(levelname)s %(name)s %(message)s")
+stream_handler.setFormatter(formatter)
+root.addHandler(stream_handler)
 
+_LOGGER = logging.getLogger(__name__)
 
-def _set_json_file(path, content):
-    full_path = os.path.join(
-        os.path.dirname(__file__),
-        f"..\\custom_components\\{DOMAIN}\\{path}"
-    )
+SOURCE_LANGUAGE = "en"
+DESTINATION_LANGUAGES = {
+    "en": "en",
+    "it": "it"
+}
 
-    data = json.dumps(content, indent=4)
-
-    with open(full_path, "w", encoding="utf-8") as json_file:
-        json_file.write(data)
-
-
-def _get_gaps(lang_name: str):
-    strings_json = _get_json_file("strings.json")
-    lang_json = _get_json_file(f"translations\\{lang_name}.json")
-
-    strings_keys = flatten(strings_json, separator=".")
-    lang_keys = flatten(lang_json, separator=".")
-
-    added_gaps = [key for key in strings_keys if key not in lang_keys]
-    removed_gaps = [key for key in lang_keys if key not in strings_keys]
-
-    if len(added_gaps) > 0:
-        for key in strings_keys:
-            if key not in lang_keys:
-                key_parts = key.split(".")
-                data_to_handle = lang_json
-                data_source = strings_json
-
-                for i in range(0, len(key_parts)):
-                    key_item = key_parts[i]
-
-                    if i == len(key_parts) - 1:
-                        data_to_handle[key_item] = f"*{data_source[key_item]}*"
-
-                    else:
-                        if key_item not in data_to_handle:
-                            data_to_handle[key_item] = {}
-                        data_to_handle = data_to_handle[key_item]
-                        data_source = data_source[key_item]
-
-        _set_json_file(f"translations\\{lang_name}.json", lang_json)
-
-    gaps = {
-        "added": added_gaps,
-        "removed": removed_gaps
-    }
-
-    return gaps
+TRANSLATION_PROVIDER = "google"
+FLAT_SEPARATOR = "."
 
 
-def _compare():
-    for lang_name in SUPPORTED_LANGUAGES:
-        gaps = _get_gaps(lang_name)
-        added_keys = gaps.get("added")
-        removed_keys = gaps.get("removed")
+class TranslationGenerator:
+    def __init__(self):
+        self._source_translations = self._get_source_translations()
 
-        if len(added_keys) + len(removed_keys) > 0:
-            print(f"Translations for '{lang_name}' is not up to date.")
+        self._destinations = DESTINATION_LANGUAGES
 
-            if len(added_keys) > 0:
-                print(f"New keys:")
-                for key in added_keys:
-                    print(f" - {key}")
+    async def initialize(self):
+        values = flatten(self._source_translations, FLAT_SEPARATOR)
+        value_keys = list(values.keys())
+        last_key = value_keys[len(value_keys) - 1]
 
-            if len(removed_keys) > 0:
-                print(f"Removed keys:")
-                for key in removed_keys:
-                    print(f" - {key}")
+        _LOGGER.info(
+            f"Process will translate {len(values)} sentences "
+            f"to {len(list(self._destinations.keys()))} languages"
+        )
+
+        for lang in self._destinations:
+            original_values = values.copy()
+            translated_data = self._get_translations(lang)
+            translated_values = flatten(translated_data, FLAT_SEPARATOR)
+
+            provider_lang = self._destinations[lang]
+            lang_cache = {}
+
+            lang_title = provider_lang.upper()
+
+            for key in original_values:
+                english_value = original_values[key]
+
+                if not isinstance(english_value, str):
+                    continue
+
+                if key in translated_values:
+                    translated_value = translated_values[key]
+
+                    _LOGGER.debug(
+                        f"Skip translation to '{lang_title}', "
+                        f"translation of '{english_value}' already exists - '{translated_value}'"
+                    )
+
+                    continue
+
+                if english_value in lang_cache:
+                    translated_value = lang_cache[english_value]
+
+                    _LOGGER.debug(
+                        f"Skip translation to '{lang_title}', "
+                        f"translation of '{english_value}' available in cache - {translated_value}"
+                    )
+
+                elif lang == SOURCE_LANGUAGE:
+                    translated_value = english_value
+
+                    _LOGGER.debug(
+                        f"Skip translation to '{lang_title}', "
+                        f"source and destination languages are the same - {translated_value}"
+                    )
+
+                else:
+                    original_english_value = english_value
+
+                    sleep_seconds = 10 if last_key == key else 0
+
+                    translated_value = ts.translate_text(
+                        english_value,
+                        translator=TRANSLATION_PROVIDER,
+                        to_language=provider_lang,
+                        sleep_seconds=sleep_seconds
+                    )
+
+                    lang_cache[english_value] = translated_value
+
+                    _LOGGER.debug(f"Translating '{original_english_value}' to {lang_title}: {translated_value}")
+
+                translated_values[key] = translated_value
+
+            translated_data = unflatten(translated_values, FLAT_SEPARATOR)
+
+            self._save_translations(lang, translated_data)
+
+    @staticmethod
+    def _get_source_translations() -> dict:
+        current_path = Path(__file__)
+        parent_directory = current_path.parents[1]
+        file_path = os.path.join(parent_directory, "custom_components", DOMAIN, "strings.json")
+
+        with open(file_path) as f:
+            data = json.load(f)
+
+            return data
+
+    @staticmethod
+    def _get_translations(lang: str):
+        current_path = Path(__file__)
+        parent_directory = current_path.parents[1]
+        file_path = os.path.join(parent_directory, "custom_components", DOMAIN, "translations", f"{lang}.json")
+
+        if os.path.exists(file_path):
+            with open(file_path) as file:
+                data = json.load(file)
+        else:
+            data = {}
+
+        return data
+
+    @staticmethod
+    def _save_translations(lang: str, data: dict):
+        current_path = Path(__file__)
+        parent_directory = current_path.parents[1]
+        file_path = os.path.join(parent_directory, "custom_components", DOMAIN, "translations", f"{lang}.json")
+
+        with open(file_path, "w+", encoding="utf-8") as file:
+            file.write(json.dumps(data, indent=4))
+
+        _LOGGER.info(f"Translation for {lang.upper()} stored")
 
 
-_compare()
+if __name__ == "__main__":
+    loop = asyncio.new_event_loop()
+
+    instance = TranslationGenerator()
+
+    try:
+        loop.run_until_complete(instance.initialize())
+
+    except KeyboardInterrupt:
+        _LOGGER.info("Aborted")
+
+    except Exception as rex:
+        _LOGGER.error(f"Error: {rex}")